+++ R/hotspot_classify_params.R
@@ -27,9 +27,12 @@
 #'   (if \code{include_self = FALSE}) values? You are unlikely to want to change
 #'   the default value.
 #' @param p_adjust_method The method to be used to adjust \emph{p}-values for
-#'   multiple comparisons. \code{NULL} (the default) uses the default method
-#'   used by \code{\link[stats]{p.adjust}}, but any of the character values in
-#'   \code{stats::p.adjust.methods} may be specified.
+#'   multiple comparisons using \code{\link[spdep]{p.adjustSP}}. \code{NULL}
+#'   (the default) uses the default method used by
+#'   \code{\link[stats]{p.adjust}} (currently \code{"holm"}), but any of the
+#'   character values in \code{stats::p.adjust.methods} may be specified. Set
+#'   this argument to \code{"none"} to return unadjusted \emph{p}-values. The
+#'   adjustment is applied separately to the cells in each time period.
 #'
 #' @return A list that can be used as the input to the \code{params} argument to
 #'   \code{\link{hotspot_classify}}.